+++ examples/data-sources/datahub_secrets/data-source.tf
@@ -1,0 +1,5 @@
+data "datahub_secrets" "all" {}
+
+output "secret_urns" {
+  value = data.datahub_secrets.all.urns
+}