+++ internal/provider/provider.go
@@ -246,9 +246,12 @@ func (p *datahubProvider) Configure(ctx context.Context, req provider.ConfigureR
 // DataSources defines the data sources implemented in the provider.
 func (p *datahubProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
+		NewConnectionsDataSource,
 		NewIngestionSourceDataSource,
+		NewIngestionSourcesDataSource,
 		NewMeDataSource,
 		NewRemoteExecutorPoolDataSource,
+		NewSecretsDataSource,
 	}
 }